+++ nextstrain/cli/__init__.py
@@ -15,7 +15,7 @@ from textwrap import dedent
 from types    import SimpleNamespace
 
 from .argparse    import HelpFormatter, register_commands, register_default_command
-from .command     import build, view, deploy, remote, shell, update, check_setup, login, logout, whoami, version
+from .command     import build, view, deploy, remote, shell, update, check_setup, login, logout, whoami, version, debugger
 from .errors      import NextstrainCliError
 from .util        import warn
 from .__version__ import __version__
@@ -73,6 +73,7 @@ def make_parser():
         logout,
         whoami,
         version,
+        debugger,
     ]
 
     register_default_command(parser)